Make the package version and state granular per chroot
Currently we fetch the latest (successful) build for each
package. This is a source of many problems in projects with multiple
chroots. The latest successful build can be submitted for a subset of
chroots, or even for none of the currently enabled chroots. Either way
it was considered successful and not ought to be built
again. Similarly with package versions.

With this change, when using `--new-packages` or `--new-versions`,
we are comparing the new packages and versions for each of the
`chroots` defined in a config. If `chroots` are not defined, we
compare it with all enabled chroots in the project.